requirements.txt
# AI Employee Silver Tier - Python Dependencies

# ===== CORE DEPENDENCIES (Bronze Tier) =====

# Watcher dependencies
watchdog==4.0.0

# Logging and utilities
python-dotenv==1.0.0

# ===== SILVER TIER DEPENDENCIES =====

# Gmail Watcher
google-api-python-client==2.100.0  # Gmail API
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# Scheduling and orchestration
schedule==1.2.0

# General API calls
requests==2.31.0

# ===== GOLD TIER DEPENDENCIES =====

# Browser automation (WhatsApp Watcher + Browser MCP)
playwright==1.40.0

# Web scraping and parsing
beautifulsoup4==4.12.0

# HTTP client for API integrations
httpx==0.25.0

# Data analysis for audits
pandas==2.1.0

# Date/time utilities
python-dateutil==2.8.2
